+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClient.java
@@ -4,9 +4,14 @@ import java.io.IOException;
 import java.net.http.HttpClient;
 import java.net.http.HttpResponse;
 import java.net.http.HttpResponse.BodyHandlers;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
-import java.util.Optional;
+import java.util.List;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
@@ -19,6 +24,7 @@ import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstant
 import com.getindata.connectors.http.internal.status.ComposeHttpStatusCodeChecker;
 import com.getindata.connectors.http.internal.status.ComposeHttpStatusCodeChecker.ComposeHttpStatusCodeCheckerConfig;
 import com.getindata.connectors.http.internal.status.HttpStatusCodeChecker;
+import static com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants.RESULT_TYPE;
 
 /**
  * An implementation of {@link PollingClient} that uses Java 11's {@link HttpClient}.
@@ -33,18 +39,24 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
 
     private final DeserializationSchema<RowData> responseBodyDecoder;
 
+    private final ObjectMapper objectMapper;
+
     private final HttpRequestFactory requestFactory;
 
     private final HttpPostRequestCallback<HttpLookupSourceRequestEntry> httpPostRequestCallback;
 
+    private final HttpLookupConfig options;
+
     public JavaNetHttpPollingClient(
             HttpClient httpClient,
             DeserializationSchema<RowData> responseBodyDecoder,
+            ObjectMapper objectMapper,
             HttpLookupConfig options,
             HttpRequestFactory requestFactory) {
 
         this.httpClient = httpClient;
         this.responseBodyDecoder = responseBodyDecoder;
+        this.objectMapper = objectMapper;
         this.requestFactory = requestFactory;
 
         this.httpPostRequestCallback = options.getHttpPostRequestCallback();
@@ -61,21 +73,22 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
                 .build();
 
         this.statusCodeChecker = new ComposeHttpStatusCodeChecker(checkerConfig);
+        this.options = options;
     }
 
     @Override
-    public Optional<RowData> pull(RowData lookupRow) {
+    public Collection<RowData> pull(RowData lookupRow) {
         try {
-            log.debug("Optional<RowData> pull with Rowdata={}.", lookupRow);
+            log.debug("Collection<RowData> pull with Rowdata={}.", lookupRow);
             return queryAndProcess(lookupRow);
         } catch (Exception e) {
             log.error("Exception during HTTP request.", e);
-            return Optional.empty();
+            return Collections.emptyList();
         }
     }
 
     // TODO Add Retry Policy And configure TimeOut from properties
-    private Optional<RowData> queryAndProcess(RowData lookupData) throws Exception {
+    private Collection<RowData> queryAndProcess(RowData lookupData) throws Exception {
 
         HttpLookupSourceRequestEntry request = requestFactory.buildLookupRequest(lookupData);
         HttpResponse<String> response = httpClient.send(
@@ -85,14 +98,14 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
         return processHttpResponse(response, request);
     }
 
-    private Optional<RowData> processHttpResponse(
+    private Collection<RowData> processHttpResponse(
             HttpResponse<String> response,
             HttpLookupSourceRequestEntry request) throws IOException {
 
         this.httpPostRequestCallback.call(response, request, "endpoint", Collections.emptyMap());
 
         if (response == null) {
-            return Optional.empty();
+            return Collections.emptyList();
         }
 
         String responseBody = response.body();
@@ -102,14 +115,14 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
                         "with Server response body [%s] ", statusCode, responseBody));
 
         if (notErrorCodeAndNotEmptyBody(responseBody, statusCode)) {
-            return Optional.ofNullable(responseBodyDecoder.deserialize(responseBody.getBytes()));
+            return deserialize(responseBody);
         } else {
             log.warn(
                 String.format("Returned Http status code was invalid or returned body was empty. "
                 + "Status Code [%s]", statusCode)
             );
 
-            return Optional.empty();
+            return Collections.emptyList();
         }
     }
 
@@ -121,5 +134,25 @@ public class JavaNetHttpPollingClient implements PollingClient<RowData> {
     @VisibleForTesting
     HttpRequestFactory getRequestFactory() {
         return this.requestFactory;
+    }
+
+    private Collection<RowData> deserialize(String responseBody) throws IOException {
+        byte[] rawBytes = responseBody.getBytes();
+        String resultType = options.getProperties().getProperty(RESULT_TYPE, "single-value");
+        if (resultType.equals("single-value")) {
+            return Collections.singletonList(responseBodyDecoder.deserialize(rawBytes));
+        } else if (resultType.equals("array")) {
+            List<JsonNode> rawObjects =
+                objectMapper.readValue(rawBytes, new TypeReference<>() {
+                });
+            List<RowData> result = new ArrayList<>(rawObjects.size());
+            for (JsonNode rawObject : rawObjects) {
+                result.add(responseBodyDecoder.deserialize(rawObject.toString().getBytes()));
+            }
+            return result;
+        } else {
+            throw new IllegalStateException(
+                String.format("Unknown lookup source result type '%'.", resultType));
+        }
     }
 }