+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientFactory.java
@@ -2,6 +2,7 @@ package com.getindata.connectors.http.internal.table.lookup;
 
 import java.net.http.HttpClient;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.table.data.RowData;
 
@@ -26,6 +27,7 @@ public class JavaNetHttpPollingClientFactory implements PollingClientFactory<Row
         return new JavaNetHttpPollingClient(
             httpClient,
             schemaDecoder,
+            new ObjectMapper(),
             options,
             requestFactory
         );