+++ src/test/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClientConnectionTest.java
@@ -1,7 +1,8 @@
 package com.getindata.connectors.http.internal.table.lookup;
 
+import java.util.Collection;
+import java.util.Iterator;
 import java.util.List;
-import java.util.Optional;
 import java.util.Properties;
 import java.util.stream.Stream;
 
@@ -52,12 +53,14 @@ import com.getindata.connectors.http.internal.table.lookup.querycreators.Generic
 import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
 import com.getindata.connectors.http.internal.utils.SerializationSchemaUtils;
 import static com.getindata.connectors.http.TestHelper.readTestFile;
+import static com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants.RESULT_TYPE;
 import static com.getindata.connectors.http.internal.table.lookup.HttpLookupTableSourceFactory.row;
 
 @ExtendWith(MockitoExtension.class)
 class JavaNetHttpPollingClientConnectionTest {
 
     private static final String SAMPLES_FOLDER = "/http/";
+    private static final String SAMPLES_FOLDER_ARRAY_RESULT = "/http-array-result/";
 
     private static final String ENDPOINT = "/service";
 
@@ -110,6 +113,7 @@ class JavaNetHttpPollingClientConnectionTest {
             HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
             "application/json"
         );
+        this.properties.setProperty(RESULT_TYPE, "single-value");
     }
 
     @AfterEach
@@ -127,11 +131,13 @@ class JavaNetHttpPollingClientConnectionTest {
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl());
 
         // WHEN
-        RowData result = pollingClient.pull(lookupRowData).orElseThrow();
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
 
         // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
 
+        assertThat(results).hasSize(1);
+        RowData result = results.iterator().next();
         assertThat(result.getArity()).isEqualTo(4);
         assertThat(result.getString(1)
             .toString()).isEqualTo("Returned HTTP message for parameter PARAM, COUNTER");
@@ -157,7 +163,7 @@ class JavaNetHttpPollingClientConnectionTest {
             );
 
         // WHEN
-        RowData result = pollingClient.pull(lookupRowData).orElseThrow();
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
 
         // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
@@ -169,6 +175,8 @@ class JavaNetHttpPollingClientConnectionTest {
             fail("Unexpected REST method.");
         }
 
+        assertThat(results).hasSize(1);
+        RowData result = results.iterator().next();
         assertThat(result.getArity()).isEqualTo(4);
         assertThat(result.getString(1)
             .toString()).isEqualTo("Returned HTTP message for parameter PARAM, COUNTER");
@@ -188,6 +196,43 @@ class JavaNetHttpPollingClientConnectionTest {
         );
     }
 
+    @Test
+    void shouldQuery200WithArrayResult() {
+        // GIVEN
+        this.stubMapping = setUpServerStubArrayResult(200);
+
+        Properties properties = new Properties();
+        properties.putAll(this.properties);
+        properties.setProperty(RESULT_TYPE, "array");
+
+        // WHEN
+        JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl(), properties);
+
+        // WHEN
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
+
+        // THEN
+        wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
+
+        assertThat(results).hasSize(2);
+
+        Iterator<RowData> iterator = results.iterator();
+
+        RowData firstResult = iterator.next();
+        assertThat(firstResult.getArity()).isEqualTo(4);
+        RowData detailsRow1 = firstResult.getRow(3, 2);
+        assertThat(detailsRow1.getBoolean(0)).isEqualTo(true); // isActive
+        RowData nestedDetailsRow1 = detailsRow1.getRow(1, 1);
+        assertThat(nestedDetailsRow1.getString(0).toString()).isEqualTo("$1,729.34");
+
+        RowData secondResult = iterator.next();
+        assertThat(secondResult.getArity()).isEqualTo(4);
+        RowData detailsRow2 = secondResult.getRow(3, 2);
+        assertThat(detailsRow2.getBoolean(0)).isEqualTo(false); // isActive
+        RowData nestedDetailsRow2 = detailsRow2.getRow(1, 1);
+        assertThat(nestedDetailsRow2.getString(0).toString()).isEqualTo("$22,001.99");
+    }
+
     @ParameterizedTest
     @MethodSource("clientErrorCodeConfig")
     void shouldHandleCodeBasedOnConfiguration(
@@ -203,10 +248,10 @@ class JavaNetHttpPollingClientConnectionTest {
         );
 
         // WHEN
-        Optional<RowData> poll = pollingClient.pull(lookupRowData);
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
 
         // THEN
-        assertThat(poll.isEmpty()).isEqualTo(isExpectedResponseEmpty);
+        assertThat(results.isEmpty()).isEqualTo(isExpectedResponseEmpty);
     }
 
     @Test
@@ -217,12 +262,12 @@ class JavaNetHttpPollingClientConnectionTest {
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl());
 
         // WHEN
-        Optional<RowData> poll = pollingClient.pull(lookupRowData);
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
 
         // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
 
-        assertThat(poll.isEmpty()).isTrue();
+        assertThat(results.isEmpty()).isTrue();
     }
 
     @Test
@@ -233,10 +278,10 @@ class JavaNetHttpPollingClientConnectionTest {
         JavaNetHttpPollingClient pollingClient = setUpPollingClient(getBaseUrl());
 
         // WHEN
-        Optional<RowData> poll = pollingClient.pull(null);
+        Collection<RowData> results = pollingClient.pull(null);
 
         // THEN
-        assertThat(poll.isEmpty()).isTrue();
+        assertThat(results.isEmpty()).isTrue();
     }
 
     @ParameterizedTest
@@ -252,10 +297,7 @@ class JavaNetHttpPollingClientConnectionTest {
         this.stubMapping = setupServerStubForBasicAuth();
 
         Properties properties = new Properties();
-        properties.setProperty(
-            HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
-            "application/json"
-        );
+        properties.putAll(this.properties);
 
         properties.setProperty(
             HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Authorization",
@@ -274,11 +316,13 @@ class JavaNetHttpPollingClientConnectionTest {
         );
 
         // WHEN
-        RowData result = pollingClient.pull(lookupRowData).orElseThrow();
+        Collection<RowData> results = pollingClient.pull(lookupRowData);
 
         // THEN
         wireMockServer.verify(RequestPatternBuilder.forCustomMatcher(stubMapping.getRequest()));
 
+        assertThat(results).hasSize(1);
+        RowData result = results.iterator().next();
         assertThat(result.getArity()).isEqualTo(4);
         assertThat(result.getString(1)
             .toString()).isEqualTo("Returned HTTP message for parameter PARAM, COUNTER");
@@ -295,13 +339,10 @@ class JavaNetHttpPollingClientConnectionTest {
     }
 
     public JavaNetHttpPollingClient setUpPollingClient(String url) {
+        return setUpPollingClient(url, properties);
+    }
 
-        Properties properties = new Properties();
-        properties.setProperty(
-            HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
-            "application/json"
-        );
-
+    public JavaNetHttpPollingClient setUpPollingClient(String url, Properties properties) {
         return setUpPollingClient(url, properties, setUpGetRequestFactory(properties));
     }
 
@@ -423,6 +464,16 @@ class JavaNetHttpPollingClientConnectionTest {
                         .withBody(readTestFile(SAMPLES_FOLDER + "HttpResult.json"))));
     }
 
+    private StubMapping setUpServerStubArrayResult(int status) {
+        return wireMockServer.stubFor(
+            get(urlEqualTo(ENDPOINT + "?id=1&uuid=2"))
+                .withHeader("Content-Type", equalTo("application/json"))
+                .willReturn(
+                    aResponse()
+                        .withStatus(status)
+                        .withBody(readTestFile(SAMPLES_FOLDER_ARRAY_RESULT + "HttpResult.json"))));
+    }
+
     private StubMapping setupServerStubForBasicAuth() {
         return wireMockServer.stubFor(get(urlEqualTo(ENDPOINT + "?id=1&uuid=2"))
                 .withHeader("Content-Type", equalTo("application/json"))
@@ -447,6 +498,8 @@ class JavaNetHttpPollingClientConnectionTest {
         properties.setProperty(
             HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX + "Content-Type",
             "application/json");
+
+        properties.setProperty(RESULT_TYPE, "single-value");
 
         return properties;
     }